second commit (finance-service fixed)

+++ services/finance-service/src/main/java/com/finance/finance/controllers/dto/CreateAccountRequest.java
@@ -12,5 +12,5 @@ public class CreateAccountRequest {
     private String name;
 
     @NotNull
-    private BigDecimal initialBalance;
+    private double initialBalance;
 }
